Add initial project setup
This commit contains the initial project setup. It will read the
temperature and humidity using a DHT-22 sensor and displays the
readings on an SSD1306 compatible display.

By default it will refresh the sensor readings every 30 seconds but a
webpage is available where you can change the refresh rate from
anywhere between 2 seconds and 2 minutes.